Fixed crashes etc. in XML handling and added "additional parameter" opt.
Restructured the project settings code to use interfaces throughout and fixed some crashes by this. Added a new option on Misc. options screen to add arbitrary options to the generated batch file to enable using parameters the tool does not yet know of. Disabled the "next" button on the "Save and Run" screen of the wizard.