Crash of save menu fixed when no file specified yet and other fixes.
Application calls save as now, if no file name is specified yet or if it doesn't exist. Next button on "Output" screen is now enabled when all relevant settings are set. When running the batch file produces some error, the animation is stopped and after the failure message the save and run screen is shown.